unify: extract `TermComparator` and rename `DefEq` to `Unifier`
